requirements/ml.txt
sentence-transformers
faiss-cpu
scikit-learn
torch